+++ src/parser/block/extension/fig.ts
@@ -26,5 +26,5 @@ export const segment: FigureParser = block(union([
 export const fig: FigureParser = block(rewrite(segment, source => {
   const bracket = (source.match(/^[^\n]*\n!?>+\s/) && source.match(/^~{3,}(?=\s*)$/gm) || [])
     .reduce((max, bracket) => bracket > max ? bracket : max, '~~') + '~';
-  return figure(source.replace(/^([^\n]+)\n([\s\S]+?)\n?$/, `${bracket}figure $1\n$2\n${bracket}`));
+  return figure(`${bracket}figure ${source}\n${bracket}`);
 }));